requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-CORS==6.0.0
requests>=2.32.3
python-dotenv>=1.0.0
python-dateutil>=2.8.2
pytesseract>=0.3.10
opencv-python>=4.8.0.74
numpy>=1.24.3,<2.0.0
Pillow>=10.0.0
Werkzeug>=3.1.0
pandas>=2.0.3,<3.0.0
openpyxl>=3.1.2
